Added spikes and movable box
Added spikes and box at request of designer.

spikes reload the scene rather than functionally killing the player.

added a few scripts for spikes, and for the box in order to push the player.

BUGS - if the player falls off the map after the rotation, it has a high chance to cause a softlock.